Feature: add turtle, message features

See merge request s10-final/S10P31A706!44

+++ backend/src/main/java/com/accio/isegye/turtle/repository/TurtleLogRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TurtleLogRepository extends JpaRepository<TurtleLog, Integer> {
+public interface TurtleLogRepository extends JpaRepository<TurtleLog, Long> {
 
 }
